+++ src/app/algorithms/page.tsx
@@ -1,4 +1,4 @@
-import BinarySearchTree from '@/components/BinarySearchTree';
+import { BinarySearchTree } from '@/components/algorithms';
 
 export default function AlgorithmsPage() {
   return <BinarySearchTree />;
