+++ src/components/algorithms/AlgorithmCard.tsx
