+++ src/components/algorithms/AlgorithmSelector.tsx
