+++ src/components/algorithms/CurrentAlgorithmHeader.tsx
