+++ src/components/algorithms/ResultsDisplay.tsx
