+++ src/components/algorithms/TreeVisualizer.tsx
