+++ src/components/react/index.ts
@@ -1,0 +1,1 @@
+export { CodeBlock } from './CodeBlock';
