+++ src/components/ui/index.ts
@@ -1,0 +1,8 @@
+export * from './alert';
+export * from './badge';
+export * from './button';
+export * from './card';
+export * from './dialog';
+export * from './separator';
+export * from './sheet';
+export * from './tabs';
